changed return type of getFPS to double

+++ src/modules/lecturesight-framesource-api/src/main/java/cv/lecturesight/framesource/FrameSource.java
@@ -43,6 +43,6 @@ public interface FrameSource {
   
   long getFrameNumber();
   
-  int getFPS();
+  double getFPS();
 
 }
